requirements.txt
Django>=3.1.7,<3.2.0
django_extensions>=3.1.1,<3.2.0
djangorestframework>=3.12.2,<3.13.0
drf-nested-routers>=0.93.3,<0.94.0
psycopg2>=2.7.5,<2.8.0
Pillow>8.1.1<8.2.0
flake8>=3.6.0,<3.7.0
django-measurement>=3.2.3,<3.3.0
django-taggit>=1.3.0,<1.4.0
django-cities-light>=3.8.1,<3.9.0
django-phonenumber-field[phonenumberslite]>=5.0.0,<5.1.0